create article configured, article tag many to many configured, tested article create

+++ database/migrations/2023_02_11_100928_article_tag.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('article_tag', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('article_id');
+            $table->unsignedInteger('article_id');
             $table->unsignedBigInteger('tag_id');
             $table->timestamps();
         });
